clean up toolz usage
