+++ conda/_vendor/toolz/__init__.py
@@ -23,3 +23,32 @@
 # updated: 2017-05-30
 __version__ = '0.8.2'
 
+
+try:
+    from cytoolz import __version__ as cytoolz_version
+    if tuple(int(x) for x in cytoolz_version.split(".")) < (0, 8, 2):
+        raise ImportError()
+    from cytoolz.itertoolz import (
+        accumulate, concat, concatv, drop, interleave, groupby, take, unique
+    )
+    from cytoolz.dicttoolz import merge
+    from cytoolz.functoolz import excepts
+except (ImportError, ValueError):
+    from .itertoolz import (
+        accumulate, concat, concatv, drop, interleave, groupby, take, unique
+    )
+    from .dicttoolz import merge
+
+    # Importing from toolz.functoolz is slow since it imports inspect.
+    # Copy the relevant part of excepts' implementation instead:
+    class excepts(object):
+        def __init__(self, exc, func, handler=lambda exc: None):
+            self.exc = exc
+            self.func = func
+            self.handler = handler
+
+        def __call__(self, *args, **kwargs):
+            try:
+                return self.func(*args, **kwargs)
+            except self.exc as e:
+                return self.handler(e)
