+++ conda/common/path.py
@@ -13,6 +13,7 @@ import subprocess
 from .compat import PY2, ensure_fs_path_encoding, on_win, string_types
 from .. import CondaError
 from .._vendor.auxlib.decorators import memoize
+from .._vendor.toolz import accumulate, concat, take
 
 try:
     # Python 3
@@ -21,11 +22,6 @@ except ImportError:  # pragma: no cover
     # Python 2
     from urllib import unquote  # NOQA
     from urlparse import urlsplit  # NOQA
-
-try:
-    from cytoolz.itertoolz import accumulate, concat, take
-except ImportError:  # pragma: no cover
-    from .._vendor.toolz.itertoolz import accumulate, concat, take
 
 log = getLogger(__name__)
