+++ conda/core/subdir_data.py
@@ -20,6 +20,7 @@ import warnings
 from .. import CondaError
 from .._vendor.auxlib.ish import dals
 from .._vendor.auxlib.logz import stringify
+from .._vendor.toolz import concat, take
 from ..base.constants import CONDA_HOMEPAGE_URL
 from ..base.context import context
 from ..common.compat import (ensure_binary, ensure_text_type, ensure_unicode, iteritems,
@@ -38,11 +39,6 @@ from ..gateways.disk.update import touch
 from ..models.channel import Channel, all_channel_urls
 from ..models.match_spec import MatchSpec
 from ..models.records import PackageRecord
-
-try:
-    from cytoolz.itertoolz import concat, take
-except ImportError:  # pragma: no cover
-    from .._vendor.toolz.itertoolz import concat, take  # NOQA
 
 try:
     import cPickle as pickle
