+++ conda/models/channel.py
@@ -8,6 +8,7 @@ from itertools import chain
 from logging import getLogger
 
 from .._vendor.boltons.setutils import IndexedSet
+from .._vendor.toolz import concat, concatv, drop
 from ..base.constants import DEFAULTS_CHANNEL_NAME, MAX_CHANNEL_PRIORITY, UNKNOWN_CHANNEL
 from ..base.context import context
 from ..common.compat import ensure_text_type, isiterable, iteritems, odict, with_metaclass
@@ -15,11 +16,6 @@ from ..common.path import is_path, win_path_backout
 from ..common.url import (Url, has_scheme, is_url, join_url, path_to_url,
                           split_conda_url_easy_parts, split_platform, split_scheme_auth_token,
                           urlparse)
-
-try:
-    from cytoolz.itertoolz import concat, concatv, drop
-except ImportError:  # pragma: no cover
-    from .._vendor.toolz.itertoolz import concat, concatv, drop  # NOQA
 
 log = getLogger(__name__)
